Added create_pool.php and create_pool.twig. Added a button in userHomePage.twig that directs to this page. Basic functionality added that if there is a user logged in, it allows them through - if there is no user logged in, it redirects them to login.php
Added functionality in login.php/login.twig, new possible parameter
"rd". If set (to any value), it indicates that the request was a
redirect from another page requiring a login. If set, it will inform the
user that they need to be logged in to view that page.

Idea for future implementation is to add a "next" parameter that takes
in a URL. If set, login should redirect the user to that page instead of
their home page.

Functionality would be that a page that requires login redirects to
/hockeypool/login?rd=1&next=/hockeypool/callingpage.php

Login could then direct there instead of home.